[CLD-7] add middleware to the endpoints
The middleware is responsible for checking the validation state of the token
It gets the Bearer token from the Authorization header, after validating the token, it gets user's email from the response, and pass the request to the next endpoint.
This means each endpoint no longer need to check authorization token and get email themselves.

Risk:
Change to frontend: currently the authorization token is passed to the backend as a parameter, e.g. /users/create?token=token
This needs to be changed to store token as the authorization header of the request.