[CDL-016]Connect Create Project to backend (#159)
Issue:
Allow the original UI shown normally and allow the user to create a project.

Solution:

In the Frontend, change the user and project services file to ensure the creating project process is working.
In the backend, refactor the projects/all endpoint in the backend to make sure the projects will show on the main page after creating.
In the backend, project_dao.py, add a function to find the owner of the project. The project name is the combination of the owner name and the project name.
Risk:
The user still can't create an encrypted project. The encryption key cannot send to the backend.

Reviewed By: Emily